+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Manage SELinux context of TFTP directory. #1997
 - Dynamically write `$tftpdir/warewulf/grub.cfg` to the configured value from `warewulf.conf`
 - Absolute paths specified with `{{ file }}` in an overlay now write to that absolute path.
+- Use opencontainers/selinux to manage SELinux in wwclient.
 
 ### Fixed